Initial commit of Arduino-Make-inator (ArMi)
Lots of known bugs.  You touch shit... it gonna break.  But you can
build a skeleton Arduino project using the two default (App Launched and
Loop).  That's it.  If you do anything else it will explode.  And you
will die.

Committing this as a binary project.  Which is lame as doing diffs is
pretty pointless.  But a few folks were having trouble cracking open the
text version of the project.  So I am posting this here for folks to
test against, then will be committing the text version of the project to
actually work with.